+++ src/SampleFormat.h
@@ -143,7 +143,7 @@ void      ClearSamples(samplePtr buffer, sampleFormat format,
                        size_t start, size_t len);
 
 void      ReverseSamples(samplePtr buffer, sampleFormat format,
-                         int start, int len);
+                         size_t start, size_t len);
 
 //
 // This must be called on startup and everytime NEW ditherers
