+++ src/Snap.h
@@ -98,7 +98,7 @@ public:
 
    static wxArrayString GetSnapLabels();
    static wxArrayString GetSnapValues();
-   static const wxString & GetSnapValue(int index);
+   static const wxString & GetSnapValue(size_t index);
    static int GetSnapIndex(const wxString & value);
 
 private:
