+++ src/effects/TruncSilence.cpp
@@ -58,7 +58,7 @@ static const wxChar *kActionStrings[kNumActions] =
 // Define keys, defaults, minimums, and maximums for the effect parameters
 //
 //     Name       Type     Key               Def         Min      Max                        Scale
-Param( DbIndex,   int,     XO("Db"),         0,          0,       Enums::NumDbChoices - 1,   1  );
+Param( DbIndex,   int,     XO("Db"),         0,          0,       (int)Enums::NumDbChoices - 1,   1  );
 Param( ActIndex,  int,     XO("Action"),     kTruncate,  0,       kNumActions - 1,           1  );
 Param( Minimum,   double,  XO("Minimum"),    0.5,        0.001,   10000.0,                   1  );
 Param( Truncate,  double,  XO("Truncate"),   0.5,        0.0,     10000.0,                   1  );
@@ -217,7 +217,7 @@ bool EffectTruncSilence::Startup()
    if (gPrefs->Exists(base))
    {
       mTruncDbChoiceIndex = gPrefs->Read(base + wxT("DbChoiceIndex"), 4L);
-      if ((mTruncDbChoiceIndex < 0) || (mTruncDbChoiceIndex >= Enums::NumDbChoices))
+      if ((mTruncDbChoiceIndex < 0) || (mTruncDbChoiceIndex >= (int)Enums::NumDbChoices))
       {  // corrupted Prefs?
          mTruncDbChoiceIndex = 4L;
       }
